+++ packages/categories/server/settings.js
@@ -8,29 +8,26 @@ Meteor.startup(function() {
 	if(sets && sets.place && sets.user) {
 
 		_.each(sets.place, function(active, name) {
-			if(active) {
-				cats.push(_.extend({}, K.schemas.cat, {
-					name: name,
-					active: active,
-					type: 'place'
-				}));
-			}
+			cats.push(_.extend({}, K.schemas.cat, {
+				name: name,
+				active: active,
+				type: 'place'
+			}));
 		});
 
 		_.each(sets.user, function(active, name) {
-			if(active) {
-				cats.push(_.extend({}, K.schemas.cat, {
-					name: name,
-					active: active,
-					type: 'user'
-				}));
-			}
+			cats.push(_.extend({}, K.schemas.cat, {
+				name: name,
+				active: active,
+				type: 'user'
+			}));
 		});
 
 		_.each(cats, function(cat) {
 
 			try {
-				var ret = Categories.upsert({
+				delete cat.rank;	//Don't touch rank in database!
+				var ret = Categories.upsert({	//updsate active status
 					name: cat.name,
 					type: cat.type
 				}, {
@@ -38,12 +35,11 @@ Meteor.startup(function() {
 				});
 
 			} catch(e) {
-				//throw new Meteor.Error(500, i18n('error_cats_exists',name) );
-				console.log( i18n('error_cats_exists',cat.name), 'type', cat.type );
+				console.warn('Cats: error loading categories from settings');
 				return null;
 			}
 		});
 
-		console.log('Cats: load categories from settings', cats.length);
+		console.log('Cats: loaded categories from settings', cats.length);
 	}
 });