+++ packages/categories/server/users.js
@@ -1,6 +1,6 @@
 
 Meteor.publish('usersByCategory', function(cat) {
-	if(this.userId) {
+	if(this.userId || K.settings.public.router.publicRoutes.usersCats) {
 		var cur = K.findUsersByCategory(cat);
 		console.log('Pub: usersByCategory', cat);
 		return cur;